+++ Libraries/UISpecialData.h
@@ -12,8 +12,8 @@ typedef std::shared_ptr<UISpecialData> UISpecialData_p;
 
 struct UISpecialData
 {
-	std::vector<std::pair<float, float>>		_ranges;
-	std::vector<float>							_i;
+	std::vector<std::pair<double, double>>		_ranges;
+	std::vector<double>							_i;
 
 	static UISpecialData_p create(){ return UISpecialData_p( new UISpecialData ); };
 };
