+++ SintezPPDefK/DefKNuton.h
@@ -20,7 +20,7 @@ private:
 	MatrixLine									createRightParts( const System & system );
 
 	NS_CORE InternalGearRatios					geInternalGearRatioValuesFromSystem( const System & system );
-	float										calcNorm( const MatrixLine& delta );
+	double										calcNorm( const MatrixLine& delta );
 	NS_CORE InternalGearRatios					solveNuton( const Jacobi& jacobian, System& system );
 
 	NS_CORE InternalGearRatios					findK( const NS_CORE Code& code, const NS_CORE InternalGearRatios& initialKValues, const NS_CORE Ratios& iTarget );
