+++ SintezPPKinCharacteristics/Function.h
@@ -20,10 +20,10 @@ private:
 	int											m_na;
 	int											m_nb;
 	int											m_nh;
-	float										m_a;
-	float										m_b;
-	float										m_c;
-	float										m_kpdSum;
+	double										m_a;
+	double										m_b;
+	double										m_c;
+	double										m_kpdSum;
 
 	Function() = delete;
 	Function( const NS_CORE N& n, const NS_CORE GearSetNumber gearSet );
@@ -36,7 +36,7 @@ public:
 
 	static int									sign( char c );
 	static Function_p							create( const NS_CORE InternalGearRatioValue& k, const NS_CORE W& w, const NS_CORE N& n, const NS_CORE GearSetNumber gearSet );
-	virtual float								calc( float ) const override;
+	virtual double								calc(double) const override;
 
 };
